refactor modalTag post
